fix(mobile): code quality improvements and type safety
- Extract shared default preferences to lib/defaults.ts (DRY fix)
- Extract shared text utilities (compactActivityText, relativeTime) to lib/text-utils.ts
- Fix result: any type in login.tsx using UserProfile type
- Fix catch-all union with any in types.ts (use Record<string, unknown>)
- Add AbortController timeout to ping() method
- Fix indentation bug in notifications.ts (sessionDeepLink)
- Improve code maintainability by removing duplication
